+++ R/get_configuration.R
@@ -72,8 +72,8 @@ get_configuration <- function(connection, projectname = NULL) {
 #' * This function is used by: [simplify_idaifield()].
 #'
 #'
-#' 
-#' 
+#'
+#'
 #'
 #' @export
 #'
@@ -101,8 +101,8 @@ get_field_inputtypes <- function(config, inputType = "all") {
   colnames(fields_mat) <- c("category", "field", "inputType")
   fields_mat[, 1] <- gsub("\\..*", "", names(fields))
   fields_mat[, 2] <- gsub(".*\\.", "", names(fields))
-  fields_mat[, 1] <- remove_config_names(fields_mat[, 1])
-  fields_mat[, 2] <- remove_config_names(fields_mat[, 2])
+  fields_mat[, 1] <- remove_config_names(fields_mat[, 1], silent = FALSE)
+  fields_mat[, 2] <- remove_config_names(fields_mat[, 2], silent = FALSE)
   fields_mat[, 3] <- unname(fields)
   return(fields_mat)
 }