+++ R/simplify_idaifield_resources.R
@@ -104,19 +104,23 @@ simplify_single_resource <- function(resource,
   # The function then gets the names of all the fields in the resource,
   # and checks if any of them contain a colon (:). If so, the
   # remove_config_names() function is applied to the list of field names
-  # to remove the portion after the colon. The resulting list of field names
+  # to remove the portion before the colon. The resulting list of field names
   # is then assigned back to the resource. If the category field of the resource
   # contains a colon, the remove_config_names() function is also applied to
-  # this field to remove the portion after the colon
+  # this field to remove the portion before the colon.
+  # The notification about duplicates is only displayed for the field names,
+  # since that is the only place where it could be relevant for further
+  # processing of the data (i.e. multiple columns with the same name in a
+  # table / data.frame.)
   list_names <- names(resource)
 
   if (any(grepl(":", list_names))) {
-    list_names <- remove_config_names(list_names)
+    list_names <- remove_config_names(list_names, silent = FALSE)
     names(resource) <- list_names
   }
 
   if (any(grepl(":", resource$category))) {
-    resource$category <- remove_config_names(resource$category)
+    resource$category <- remove_config_names(resource$category, silent = TRUE)
   }